riscv: linker: workaround size in addresses
In the RISC-V code model medany, only accesses around
PC +/- 2GB are allowed. This means that our abuse of
symbols as sizes, cause problems with linker relocations.

GCC does a trick where it relaxes these accesses to absolute
accesses, but this is dubious from a standards standpoint
and should not be relied upon.

Signed-off-by: Robin Kastberg <robin.kastberg@iar.com>